separated helper functions and cleaned up files
